AG-11543 - Abstract getInterfaceDocumentationModel methods
